style(lockfile): move the serialize_yaml tests to an external module
